# decrease lock range