before this pr:

```
BenchmarkRdmSelect-4          	50000000	        30.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4           	100000000	        14.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4          	100000000	        21.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkServeHTTP-4          	  200000	      6635 ns/op	   33712 B/op	      14 allocs/op
BenchmarkProxy-4              	  200000	      6160 ns/op	   33664 B/op	      12 allocs/op
BenchmarkRouterGenericURL-4   	10000000	       126 ns/op	      48 B/op	       2 allocs/op
BenchmarkIncr-4               	10000000	       139 ns/op	       0 B/op	       0 allocs/op
BenchmarkQueryStatus-4        	100000000	        22.7 ns/op	       0 B/op	       0 allocs/op
```

this pr:

```
BenchmarkRdmSelect-4          	50000000	        30.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkRRSelect-4           	100000000	        14.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkWRRSelect-4          	100000000	        21.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkServeHTTP-4          	  200000	      6361 ns/op	   33712 B/op	      14 allocs/op
BenchmarkProxy-4              	  200000	      6008 ns/op	   33664 B/op	      12 allocs/op
BenchmarkRouterGenericURL-4   	10000000	       121 ns/op	      48 B/op	       2 allocs/op
BenchmarkIncr-4               	20000000	        88.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkQueryStatus-4        	100000000	        23.1 ns/op	       0 B/op	       0 allocs/op
```