@jiajunhuang 2018-01-22
It speed up `Incr` operation.